# backend/api/openapi.yaml
# Documentation only; nothing is generated from this file. Keep it in step with
# the DTOs in internal/adapters/http/*_handler.go.
openapi: 3.0.3
info:
  title: Brag Document API
  version: 0.1.0
  description: |
    All routes except /healthz, /readyz and /metrics require a Supabase access
    token as a Bearer token. The first authenticated request provisions the user:
    an invitation for the token's email joins that tenant as `member`, otherwise
    a new tenant is created with the user as `admin`.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /me:
    get:
      summary: The authenticated caller and their tenant
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Me" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /documents:
    get:
      summary: List the caller's documents (owned and shared, archived included)
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DocumentList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create a document owned by the caller
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateDocument" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Document" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /documents/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      summary: Update title, description or state (owner only)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateDocument" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Document" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      summary: Delete a document (owner only)
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /documents/{id}/logs:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Filtered, sorted, paged logs of a document (owner only)
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Substring of name or description, case-insensitive }
        - { name: tag, in: query, schema: { type: array, items: { type: string } }, explode: true, description: Any-of }
        - { name: status, in: query, schema: { type: array, items: { $ref: "#/components/schemas/LogStatus" } }, explode: true }
        - { name: impact, in: query, schema: { type: array, items: { $ref: "#/components/schemas/Impact" } }, explode: true }
        - { name: from, in: query, schema: { type: string, format: date }, description: created_at on or after, UTC }
        - { name: to, in: query, schema: { type: string, format: date }, description: created_at on or before, UTC }
        - { name: domain, in: query, schema: { type: string }, description: Link host or any subdomain of it }
        - { name: sort, in: query, schema: { type: string, enum: [created_at, -created_at, name, -name, impact, -impact, status, -status], default: -created_at } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/LogList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    post:
      summary: Create a log; the impact statement is extracted from the text (PRD-0007)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateLog" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Log" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /documents/{id}/logs/{logId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: logId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      summary: Update a log; re-extracts the impact statement when name or description changed
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateLog" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Log" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
    delete:
      summary: Delete a log
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /documents/{id}/example-logs:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Delete the document's example logs
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /tags:
    get:
      summary: The tenant's tag vocabulary, sorted
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, required: [tags], properties: { tags: { type: array, items: { type: string } } } } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /tenant/members:
    get:
      summary: List tenant members (admin only)
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Member" } } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenant/members/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Remove a member (admin only; 422 for yourself, 409 if they still own documents)
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /tenant/invitations:
    get:
      summary: List pending invitations (admin only)
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Invitation" } } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: Invite an email to the tenant (admin only; 409 if already a member or invited)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateInvitation" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Invitation" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /tenant/invitations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Revoke an invitation (admin only)
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Unauthorized: { description: Missing or invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Forbidden: { description: Not allowed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    NotFound: { description: Resource not found (also for other tenants' rows), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Conflict: { description: Conflict with current state, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    TooLarge: { description: Request body exceeds 1 MiB, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Unprocessable: { description: Validation failed; `fields` names the offending inputs, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
    Tenant:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
    Me:
      type: object
      required: [id, email, display_name, role, tenant]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [admin, member] }
        tenant: { $ref: "#/components/schemas/Tenant" }
    Document:
      type: object
      required: [id, owner_id, title, description, state, created_at, updated_at, log_count, last_log_at]
      properties:
        id: { type: string, format: uuid }
        owner_id: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        state: { type: string, enum: [active, archived] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        log_count: { type: integer, description: Logs excluding examples }
        last_log_at: { type: string, format: date-time, nullable: true, description: Newest created_at, null without logs }
    DocumentList:
      type: object
      required: [owned, shared]
      properties:
        owned: { type: array, items: { $ref: "#/components/schemas/Document" } }
        shared: { type: array, items: { $ref: "#/components/schemas/Document" }, description: Always empty in v1 }
    CreateDocument:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
    UpdateDocument:
      type: object
      description: Absent fields are unchanged.
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        state: { type: string, enum: [active, archived] }
    Member:
      type: object
      required: [id, email, display_name, role, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [admin, member] }
        created_at: { type: string, format: date-time }
    Invitation:
      type: object
      required: [id, email, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
    CreateInvitation:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    Error:
      type: object
      required: [error, message, request_id]
      properties:
        error: { type: string, enum: [unauthorized, too_large, validation, not_found, forbidden, conflict, timeout, unavailable, internal] }
        message: { type: string }
        request_id: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
    Impact: { type: string, enum: [low, medium, high, critical] }
    LogStatus: { type: string, enum: [idea, in_progress, done, dropped] }
    Link:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri, description: http or https only }
        label: { type: string, maxLength: 100 }
    Log:
      type: object
      required: [id, document_id, name, description, impact, impact_statement, status, is_example, tags, links, created_at, created_by, updated_at, updated_by]
      properties:
        id: { type: string, format: uuid }
        document_id: { type: string, format: uuid }
        name: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 20000, description: Markdown }
        impact: { $ref: "#/components/schemas/Impact" }
        impact_statement: { type: string, nullable: true, maxLength: 280, description: 'null: not checked; "": none found (PRD-0007)' }
        status: { $ref: "#/components/schemas/LogStatus" }
        is_example: { type: boolean }
        tags: { type: array, items: { type: string } }
        links: { type: array, items: { $ref: "#/components/schemas/Link" } }
        created_at: { type: string, format: date-time }
        created_by: { type: string, format: uuid }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string, format: uuid }
    LogList:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Log" } }
        total: { type: integer }
    CreateLog:
      type: object
      required: [name, impact]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 20000 }
        impact: { $ref: "#/components/schemas/Impact" }
        status: { $ref: "#/components/schemas/LogStatus" }
        tags: { type: array, maxItems: 20, items: { type: string, maxLength: 50 } }
        links: { type: array, maxItems: 20, items: { $ref: "#/components/schemas/Link" } }
        created_at: { type: string, format: date-time, description: Defaults to now; set to back-date. Not more than 24 h in the future }
    UpdateLog:
      type: object
      description: Absent fields are unchanged; tags and links replace the whole set when present.
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 20000 }
        impact: { $ref: "#/components/schemas/Impact" }
        status: { $ref: "#/components/schemas/LogStatus" }
        tags: { type: array, maxItems: 20, items: { type: string, maxLength: 50 } }
        links: { type: array, maxItems: 20, items: { $ref: "#/components/schemas/Link" } }
        created_at: { type: string, format: date-time, description: Not more than 24 h in the future }